refactor: simplify logger imports and remove unnecessary index file
- Remove logger index.ts barrel file
- Update all imports to use Logger.ts directly
- Add static factory method Logger.create() for consistency
- Export singleton logger instance from Logger.ts
- Eliminate unnecessary indirection for cleaner imports
