+++ src/components/contact/__tests__/ContactForm.unit.test.tsx
@@ -3,7 +3,7 @@ import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
 import { ContactForm } from '../ContactForm';
-import { logger, InMemoryLogger } from '@/lib/logger';
+import { logger, InMemoryLogger } from '@/lib/logger/Logger';
 
 // Mock the reCAPTCHA hook
 const mockExecuteRecaptcha = vi.fn();
