--- src/lib/logger/index.ts
@@ -1,2 +1,0 @@
-export { Logger, InMemoryLogger, logger } from './Logger';
-export type { LogEntry } from './Logger';
