+++ tests/setup/logger.ts
@@ -1,5 +1,5 @@
 import { beforeEach, afterEach } from 'vitest';
-import { logger, InMemoryLogger } from '@/lib/logger';
+import { logger, InMemoryLogger } from '@/lib/logger/Logger';
 
 // Automatically clear logger before and after each test
 beforeEach(() => {
