+++ tests/utils/redisCleanup.ts
@@ -1,7 +1,7 @@
 import { Redis } from '@upstash/redis';
 import { config } from 'dotenv';
 import { join } from 'path';
-import { logger } from '@/lib/logger';
+import { logger } from '@/lib/logger/Logger';
 
 // Load environment variables from .env.local for Redis access
 config({ path: join(process.cwd(), '.env.local') });
